Quote identifier names when --no-color is specified
The output should be "unused variable 'helper'"  instead of "unused variable helper" when --no-color is specified. This makes the output unambiguous and more readable, and has the big benefit of making the variable name easier to parse (important for IDEs and other tools that need to parse the output).

The code in `format.lua` may need refactoring.