Revert 81e9b7b
When propogating values, it is not safe to mark an item as visited, then
figure the propogation stops and mark value as live at previous item,
as the ending item could be approached using another jump,
and, as it is now marked as visited, that second approach won't trigger
it being marked as live.

Simpliest workaround is to revert for now. It may be possible to fix this by
remembering where a value was marked as live and marking value as live before
checking if current item is visited.