Skip directories can't be scanned instead of aborting the whole check.
Assumed we have directory structure like this:
.
├── A
│   └── B
│   └── C
└── D

If directory B could not be opened, previously luacheck will abort.
Thus we have to run luacheck separately for directory C and D, and more.

With this patch, luacheck will continue to scan directory C and D,
and report the I/O error for directory B.