Better help message and behaviour with lfs missing
Instead of hiding options that are disabled without lfs,
show them and add a warning mentioning lfs.

This way it's clear that the option exists and it's clear
why exactly it's not working.

Do not error when trying to use `--[no-]cache`
without lfs, ignore it. This allows scripts
always pass `--cache` when running luacheck, without
worrying if lfs is installed or not.