Avoid working with unique but useless option tables in filter
The way runner calls filter, option stacks are already prepared
for each file. Filter still puts root option tables in the beginning
of the stack, which messes with caching since it relies on
table identities, and per-file option table roots are unique (per-file).

As a workaround, detect option-less tables and avoid them in that
specific state. In the future runner should use filter directly
and the old interface/luacheck.process_reports should be removed.