Add Dockerfile implementation for luacheck
A simple Dockerfile pulls in the repository and builds the repo inside the
container. The container hence will always contain the latest version of the
master branch. The container can be used to check any lua source folder without
installing any of the dependencies on the host operating system, making it an
ideal tool for CI integration.

To build it locally:
docker build -t luacheck:latest .
To run it locally:
docker run -it --rm -v '$(pwd):/workdir' -w /workdir luacheck:latest