docs!: Follow semver guidelines, next release will be v1.x.y
BREAKING CHANGE: According to the SemVer spec the v0.x.y release scheme
signals that we do not consider the public facing API and usage stable.
That isn't true at all for this project, it is quite stable. Any changes
we do make are carefully considered not to break existing usage.
Additional warnings may be added, but the way it is implemented and used
is expected to remain the same.

I probably should have done this bump earlier and there is nothing magic
about this release. This is only an announcement that things are
considered stable and that we'll bump version numbers in future releases
roughly corresponding to SemVer spec.